sdk/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-sdk"
version = "0.1.0"
description = "agentOS Skill Development Kit"
requires-python = ">=3.10"
dependencies = ["pyyaml>=6.0"]

[project.scripts]
agent-sdk = "agentos.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["agentos"]
